+++ src/test/java/gr/jchrist/gitextender/GitTestUtil.java
@@ -4,30 +4,22 @@ import com.intellij.notification.Notification;
 import com.intellij.notification.NotificationType;
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.components.ServiceManager;
-import com.intellij.openapi.extensions.Extensions;
 import com.intellij.openapi.project.Project;
 import com.intellij.openapi.vcs.ProjectLevelVcsManager;
 import com.intellij.openapi.vcs.impl.ProjectLevelVcsManagerImpl;
 import com.intellij.openapi.vfs.LocalFileSystem;
 import com.intellij.openapi.vfs.VirtualFile;
-import com.intellij.util.containers.ContainerUtil;
-import com.intellij.vcs.log.VcsLogProvider;
-import com.intellij.vcs.log.impl.VcsLogManager;
 import git4idea.GitUtil;
 import git4idea.GitVcs;
-import git4idea.log.GitLogProvider;
 import git4idea.repo.GitRepository;
 import org.jetbrains.annotations.NotNull;
 import org.picocontainer.MutablePicoContainer;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.List;
 
 import static gr.jchrist.gitextender.GitExecutor.*;
-import static junit.framework.Assert.assertNotNull;
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
+import static org.assertj.core.api.Assertions.assertThat;
 
 public class GitTestUtil {
 
@@ -56,7 +48,7 @@ public class GitTestUtil {
         cloneRepo(remotePath, root, false);
 
         VirtualFile gitDir = LocalFileSystem.getInstance().refreshAndFindFileByIoFile(new File(root, GitUtil.DOT_GIT));
-        assertNotNull(gitDir);
+        assertThat(gitDir).isNotNull();
         GitRepository repo = registerRepo(project, root);
 
         cd(root);
@@ -72,9 +64,9 @@ public class GitTestUtil {
         ProjectLevelVcsManagerImpl vcsManager = (ProjectLevelVcsManagerImpl)ProjectLevelVcsManager.getInstance(project);
         vcsManager.setDirectoryMapping(root, GitVcs.NAME);
         VirtualFile file = LocalFileSystem.getInstance().findFileByIoFile(new File(root));
-        assertFalse(vcsManager.getAllVcsRoots().length == 0);
+        assertThat(vcsManager.getAllVcsRoots().length).isNotZero();
         GitRepository repository = GitUtil.getRepositoryManager(project).getRepositoryForRoot(file);
-        assertNotNull("Couldn't find repository for root " + root, repository);
+        assertThat(repository).as("Couldn't find repository for root " + root).isNotNull();
         return repository;
     }
 
@@ -95,9 +87,11 @@ public class GitTestUtil {
                                                   @NotNull String title,
                                                   @NotNull String content,
                                                   @NotNull Notification actual) {
-        assertEquals("Incorrect notification type: " + tos(actual), type, actual.getType());
-        assertEquals("Incorrect notification title: " + tos(actual), title, actual.getTitle());
-        assertEquals("Incorrect notification content: " + tos(actual), cleanupForAssertion(content), cleanupForAssertion(actual.getContent()));
+        assertThat(actual.getType()).as("Incorrect notification type: " + tos(actual)).isEqualTo(type);
+        assertThat(actual.getTitle()).as("Incorrect notification title: " + tos(actual)).isEqualTo(title);
+        assertThat(cleanupForAssertion(actual.getContent()))
+                .as("Incorrect notification content: " + tos(actual))
+                .isEqualTo(cleanupForAssertion(content));
         return actual;
     }